+++ CHANGELOG.md
@@ -6,7 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.0.1] - 2014-05-31
+### Added
+- Package `dto` created to hold all payload objects
+- Package `api.Client` to call Clockfy's API
+- Command `clockify-cli workspaces` created, with options to format the output
+
+## [0.0.1] - 2019-03-03
 ### Added
 - This CHANGELOG file to hopefully serve as an evolving example of a
   standardized open source project CHANGELOG.